+++ exasol/analytics/query_handler/execution_graph.py
