+++ exasol/analytics/query_handler/graph/parameter.py
