+++ exasol/analytics/query_handler/graph/result.py
