+++ exasol/analytics/query_handler/graph/stage/sql/execution/find_object_proxies.py
