+++ exasol/analytics/query_handler/graph/stage/sql/execution/input.py
@@ -2,9 +2,8 @@ import dataclasses
 
 from exasol_bucketfs_utils_python.abstract_bucketfs_location import AbstractBucketFSLocation
 
-from exasol_machine_learning_library.execution.sql_stage_graph.sql_stage_graph import SQLStageGraph
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_input_output import \
-    SQLStageInputOutput
+from exasol.analytics.query_handler.graph.stage.sql.sql_stage_graph import SQLStageGraph
+from exasol.analytics.query_handler.graph.stage.sql.execution.input_output import SQLStageInputOutput
 
 
 @dataclasses.dataclass(frozen=True, eq=True)
