+++ exasol/analytics/query_handler/graph/stage/sql/execution/object_proxy_reference_counter.py
