+++ exasol/analytics/query_handler/graph/stage/sql/execution/query_handler_state.py
@@ -6,16 +6,12 @@ from exasol.analytics.query_handler.context.scope import ScopeQueryHandlerContex
 from exasol.analytics.query_handler.query_handler import QueryHandler
 from exasol.analytics.query_handler.result import Continue, Finish
 
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.find_object_proxies import find_object_proxies
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.object_proxy_reference_counting_bag import \
-    ObjectProxyReferenceCountingBag
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_graph_execution_input import \
-    SQLStageGraphExecutionInput
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_input_output import \
-    SQLStageInputOutput
-from exasol_machine_learning_library.execution.stage_graph.sql_stage_train_query_handler import \
-    SQLStageTrainQueryHandlerInput
-from exasol_machine_learning_library.execution.stage_graph.sql_stage import SQLStage
+from exasol.analytics.query_handler.graph.stage.sql.execution.find_object_proxies import find_object_proxies
+from exasol.analytics.query_handler.graph.stage.sql.execution.object_proxy_reference_counting_bag import ObjectProxyReferenceCountingBag
+from exasol.analytics.query_handler.graph.stage.sql.execution.input import SQLStageGraphExecutionInput
+from exasol.analytics.query_handler.graph.stage.sql.execution.input_output import SQLStageInputOutput
+from exasol.analytics.query_handler.graph.stage.sql.sql_stage_train_query_handler import SQLStageTrainQueryHandlerInput
+from exasol.analytics.query_handler.graph.stage.sql.sql_stage import SQLStage
 
 
 class ResultHandlerReturnValue(enum.Enum):
