+++ exasol/analytics/query_handler/graph/stage/sql/sql_stage_train_query_handler.py
@@ -5,8 +5,7 @@ from typing import List, Sized
 from exasol.analytics.query_handler.query_handler import QueryHandler
 from exasol_bucketfs_utils_python.abstract_bucketfs_location import AbstractBucketFSLocation
 
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_input_output import \
-    SQLStageInputOutput
+from exasol.analytics.query_handler.graph.stage.sql.execution.input_output import SQLStageInputOutput
 
 
 def is_empty(obj: Sized):
