+++ exasol/analytics/query_handler/graph/stage/stage.py
@@ -1,7 +1,8 @@
 from abc import ABC
 from typing import Generic, TypeVar
 
-from exasol_machine_learning_library.execution.trainable_estimators import Parameter, Result
+from exasol.analytics.query_handler.graph.parameter import Parameter
+from exasol.analytics.query_handler.graph.result import Result
 
 ParameterType = TypeVar("ParameterType", bound=Parameter)
 ResultType = TypeVar("ResultType", bound=Result)
