--- exasol_machine_learning_library/execution/sql_stage_graph/sql_stage_graph.py
@@ -1,4 +1,0 @@
-from exasol_machine_learning_library.execution.execution_graph import ExecutionGraph
-from exasol_machine_learning_library.execution.stage_graph.sql_stage import SQLStage
-
-SQLStageGraph = ExecutionGraph[SQLStage]
