--- exasol_machine_learning_library/execution/stage_graph/stage_graph.py
@@ -1,4 +1,0 @@
-from exasol_machine_learning_library.execution.execution_graph import ExecutionGraph
-from exasol_machine_learning_library.execution.stage_graph.stage import Stage
-
-StageGraph = ExecutionGraph[Stage]
