--- exasol_machine_learning_library/execution/trainable_estimators/__init__.py
@@ -1,2 +1,0 @@
-from .result import Result
-from .parameter import Parameter
