+++ tests/unit_tests/sql_stage_graph/stage_graph_execution_query_handler/test_state_diamond.py
@@ -5,11 +5,9 @@ import pytest
 from exasol.analytics.query_handler.query_handler import QueryHandler
 from exasol.analytics.query_handler.result import Finish, Continue
 
-from exasol_machine_learning_library.execution.sql_stage_graph.sql_stage_graph import SQLStageGraph
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_graph_execution_query_handler_state import \
-    ResultHandlerReturnValue
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.sql_stage_input_output import \
-    SQLStageInputOutput
+from exasol.analytics.query_handler.graph.stage.sql.sql_stage_graph import SQLStageGraph
+from exasol.analytics.query_handler.graph.stage.sql.execution.query_handler_state import ResultHandlerReturnValue
+from exasol.analytics.query_handler.graph.stage.sql.execution.input_output import SQLStageInputOutput
 from tests.unit_tests.sql_stage_graph.stage_graph_execution_query_handler.assert_helper import \
     assert_reference_counting_bag_creation, assert_stage_train_query_handler_created, assert_stage_not_called, \
     assert_reference_counting_bag_not_called, assert_release_on_query_handler_context_for_stage, \
