+++ tests/unit_tests/sql_stage_graph/test_data_partition.py
@@ -12,8 +12,8 @@ from exasol.analytics.schema import (
     ColumnType,
 )
 
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.data_partition import DataPartition
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.dependency import Dependency
+from exasol.analytics.query_handler.graph.stage.sql.execution.data_partition import DataPartition
+from exasol.analytics.query_handler.graph.stage.sql.execution.dependency import Dependency
 
 
 class TestEnum(Enum):
