+++ tests/unit_tests/sql_stage_graph/test_object_proxy_reference_counter.py
@@ -6,8 +6,7 @@ import pytest
 from exasol.analytics.query_handler.context.proxy.object_proxy import ObjectProxy
 from exasol.analytics.query_handler.context.scope import ScopeQueryHandlerContext
 
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.object_proxy_reference_counter import \
-    ObjectProxyReferenceCounter, ReferenceCounterStatus
+from exasol.analytics.query_handler.graph.stage.sql.execution.object_proxy_reference_counter import ObjectProxyReferenceCounter, ReferenceCounterStatus
 from tests.unit_tests.sql_stage_graph.mock_cast import mock_cast
 
 MockScopeQueryHandlerContext = Union[ScopeQueryHandlerContext, MagicMock]
