+++ tests/unit_tests/sql_stage_graph/test_object_proxy_reference_counting_bag_with_mocks.py
@@ -6,10 +6,8 @@ import pytest
 from exasol.analytics.query_handler.context.proxy.object_proxy import ObjectProxy
 from exasol.analytics.query_handler.context.scope import ScopeQueryHandlerContext
 
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.object_proxy_reference_counter import \
-    ObjectProxyReferenceCounter, ReferenceCounterStatus
-from exasol_machine_learning_library.execution.sql_stage_graph_execution.object_proxy_reference_counting_bag import \
-    ObjectProxyReferenceCountingBag, ObjectProxyReferenceCounterFactory
+from exasol.analytics.query_handler.graph.stage.sql.execution.object_proxy_reference_counter import ObjectProxyReferenceCounter, ReferenceCounterStatus
+from exasol.analytics.query_handler.graph.stage.sql.execution.object_proxy_reference_counting_bag import ObjectProxyReferenceCountingBag, ObjectProxyReferenceCounterFactory
 from tests.unit_tests.sql_stage_graph.mock_cast import mock_cast
 
 MockObjectProxyReferenceCounter = Union[ObjectProxyReferenceCounter, MagicMock]
